Improved and changed the make_buffer function. It now returns the line as its first argument always, otherwise nil, errmsg
I updated scripts that use make_buffer.

When Lua 5.2 comes out, we may be able to use a function I've added
called stdnse.lines (commented for now). It cannot be used yet because
we cannot yield inside generic for loop. The function is built for
generic for loops. It will work as such:
for line in stdnse.lines(socket) do
  ...
end