Set sa.sndd_8022_ethertype to 0 in eth-ndd.c.
This file is used for Ethernet handles on AIX. We had a report from
Peter O'Gorman that setting this member to 0 rather than ETH_TYPE_IP was
necessary to avoid an EEXIST error on bind.

http://seclists.org/nmap-dev/2011/q4/349
http://seclists.org/nmap-dev/2011/q4/353
http://seclists.org/nmap-dev/2011/q4/att-353/aix.patch
https://code.google.com/p/libdnet/issues/detail?id=23

However Ben Lentz more recently tried the patch and said that he still
got an EEXIST on bind.

http://seclists.org/nmap-dev/2012/q4/327
http://seclists.org/nmap-dev/2012/q4/395

So who knows what the right thing to do here is. The messages from IBM
seem to say that setting ETH_TYPE_IP is definitely wrong though.