Fix /author/.../feed/ capture in http-wordpress-users.nse.
Patch by Duarte.
https://seclists.org/nmap-dev/2019/q3/17

The regular expression for the case when the server returns a 200 HTTP
status is too open, it needs to be more specific.

When there is a link similar to the feed link that occurs before the
feed link, the expression will include everything between that initial
link and the feed link.

To fix this, the group that matches/captures the author username will
only do so until it finds a forward slash.